FM-41: Split up state
