+++ fendermint/vm/interpreter/src/fvm/mod.rs
@@ -12,7 +12,7 @@ pub use check::FvmCheckRet;
 pub use exec::FvmApplyRet;
 pub use fendermint_vm_message::query::FvmQuery;
 pub use query::FvmQueryRet;
-pub use state::{FvmCheckState, FvmQueryState, FvmState};
+pub use state::{FvmCheckState, FvmExecState, FvmQueryState};
 
 pub type FvmMessage = fvm_shared::message::Message;
